chore: converted to disk usage in gb as a float

+++ crates/krill-tui/src/app.rs
@@ -45,8 +45,8 @@ pub struct App {
     pub cpu_usage: f32,
     pub memory_used_mb: u64,
     pub memory_total_mb: u64,
-    pub disk_usage_gb: u64,
-    pub disk_total_gb: u64,
+    pub disk_usage_gb: f32,
+    pub disk_total_gb: f32,
 }
 
 impl App {
@@ -67,8 +67,8 @@ impl App {
             cpu_usage: 0.0,
             memory_used_mb: 0,
             memory_total_mb: 0,
-            disk_usage_gb: 0,
-            disk_total_gb: 0,
+            disk_usage_gb: 0.0,
+            disk_total_gb: 0.0,
         }
     }
